Add the actual CMake build
These are the files that take care of getting the code for FSON,
configuring, and installing the library and modules.  Retrieving the
source is gone in one of two ways.  The first is to simply give CMake
the path to an existing tree using the `FSON_DIR` variable.
Alternatively, this can use [FetchContent][FetchContent] to retrieve the
latest source at configure time.  The configuration step will attempt to
use a system installed version of FetchContent if available, but this
does include a fallback copy if necessary.

Note, this only deals with building the library and does not yet build
the tests.

[FetchContent]: https://cmake.org/cmake/help/latest/module/FetchContent.html